+++ sites/geohub/src/lib/server/helpers/getDatasetById.ts
@@ -53,8 +53,8 @@ export const getDatasetById = async (id: string, is_superuser: boolean, user_ema
               x.access_level,
               x.createdat, 
               x.created_user,
-              x.updatedat,
-              x.updated_user,
+              CASE WHEN x.updatedat is not null THEN x.updatedat ELSE x.createdat END AS updatedat,
+			        CASE WHEN x.updated_user is not null THEN x.updated_user ELSE x.created_user END AS updated_user,
               y.tags,
               CASE WHEN z.no_stars is not null THEN z.no_stars ELSE 0 END as no_stars,
               ${